requirements.txt
# requirements.txt - Ballers App Dependencies

# Core Streamlit
streamlit>=1.28.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0

# Google APIs
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0

# Visualization
plotly>=5.17.0

# Date/Time
python-dateutil>=2.8.2

# Environment Management
python-dotenv>=1.0.0

# UI Components
streamlit-option-menu>=0.3.6

# Utilities
unidecode>=1.3.7

#  Export Functionality
reportlab>=4.0.0

# Optional: Enhanced PDF features
Pillow>=10.0.0  # For image processing in PDFs
matplotlib>=3.7.0  # For chart generation in PDFs

# Testing Framework
pytest>=7.4.0
pytest-cov>=4.1.0  # Test coverage
pytest-mock>=3.11.0  # Mocking utilities
